Rename alignment utility and update imports
