+++ Crop_and_mask.py
@@ -1,4 +1,4 @@
-from src.utils.alignmengt import crop_faces, calc_alignment_coefficients, crop_faces_from_image
+from src.utils.alignment import crop_faces, calc_alignment_coefficients, crop_faces_from_image
 from PIL import Image
 import argparse, os, sys, glob
 import cv2
